Add auth state checker and fix topics page on mobile and add support for PWA
