+++ src/components/Routes/NotFound.jsx
@@ -2,7 +2,10 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 const NotFound = () => (
-  <div className="full-height-bg card-container primary-grad">
+  <div
+    className="full-height-bg card-container primary-grad"
+    style={{ display: 'flex', 'justify-content': 'center', alignItems: 'center' }}
+  >
     <Link to="/">
       <h1 className="is-family-secondary" style={{ 'font-size': '80pt' }}>
         404.
